+++ db/migrate/20210629114219_change_types_orders.rb
@@ -1,7 +1,7 @@
 class ChangeTypesOrders < ActiveRecord::Migration[6.1]
   def change
-    change_column :orders, :room_id, :bigint
-    change_column :orders, :user_id, :bigint
+    change_column :orders, :room_id, :integer
+    change_column :orders, :user_id, :integer
     change_column :orders, :status, :integer
   end
 end
