+++ db/migrate/20210709070848_add_room_id_to_images.rb
@@ -1,5 +1,5 @@
 class AddRoomIdToImages < ActiveRecord::Migration[6.1]
   def change
-    add_column :images, :room_id, :bigint
+    add_column :images, :room_id, :integer
   end
 end
